+++ include/World.h
@@ -23,10 +23,10 @@ public:
   void setLight(const Light &light);
   void setDefault();
   Intersections intersect_world(const Ray &ray);
-  glm::vec3 shade_hit(const Computation &computation);
-  glm::vec3 color_at(const Ray &ray);
+  glm::dvec3 shade_hit(const Computation &computation);
+  glm::dvec3 color_at(const Ray &ray);
   void changeSphereMaterial(size_t sphereIndex, const Material &newMaterial);
-  bool isShadowed(const glm::vec4 &point);
+  bool isShadowed(const glm::dvec4 &point);
 };
 
 #endif //RAYTRACERV2_WORLD_H
