+++ src/Canvas.cpp
@@ -8,16 +8,16 @@
 Canvas::Canvas(int windowWidth, int windowHeight) : window_width_(
     windowWidth), window_height_(windowHeight) {
   for (int i = 0; i < windowHeight; i++) {
-    std::vector<glm::vec3> temp_vector(windowWidth, glm::ivec3(0));
+    std::vector<glm::dvec3> temp_vector(windowWidth, glm::dvec3(0));
     image_.push_back(temp_vector);
   }
 }
 
-void Canvas::writePixel(const int &y, const int &x, const glm::vec3 &color) {
+void Canvas::writePixel(const int &y, const int &x, const glm::dvec3 &color) {
   image_[x][y] = color;
 }
 
-const glm::vec3 &Canvas::pixelAt(const int &y, const int &x) const {
+const glm::dvec3 &Canvas::pixelAt(const int &y, const int &x) const {
   return image_[x][y];
 }
 
@@ -27,10 +27,10 @@ void Canvas::canvas_to_ppm() {
   myFile << "P3\n" << window_width_ << " " << window_height_ << "\n" << "255" << "\n";
   for (int i = 0; i < window_height_; i++) {
     for (int j = 0; j < window_width_; j++) {
-      auto temp = image_[i][j] * 256.0f;
-      int x = static_cast<int>(std::clamp(temp.x, 0.0f, 255.0f));
-      int y = static_cast<int>(std::clamp(temp.y, 0.0f, 255.0f));
-      int z = static_cast<int>(std::clamp(temp.z, 0.0f, 255.0f));
+      auto temp = image_[i][j] * 256.0;
+      int x = static_cast<int>(std::clamp(temp.x, 0.0, 255.0));
+      int y = static_cast<int>(std::clamp(temp.y, 0.0, 255.0));
+      int z = static_cast<int>(std::clamp(temp.z, 0.0, 255.0));
       myFile << x << " " << y << " " << z << " ";
     }
     myFile << "\n";
